requirements.txt
torch>=2.0.0
numpy
Pillow
aiohttp
